# Reserve memory before projecting rows