@yingsu00 2020-09-29
> @yingsu00 The new `BlockBuilder#reserve` method is used only in one place, in GeneratedPageProjection, after a call to `BlockBuilder#newBlockBuilderLike`. It would be clearer to pass a parameter to `newBlockBuilderLike` that specifies the expected number of positions, similar to `expectedEntries` parameter in `com.facebook.presto.common.type.Type#createBlockBuilder(com.facebook.presto.common.block.BlockBuilderStatus, int)`.
> 
> ```
> blockBuilders = blockBuilders.stream().map(blockBuilder -> blockBuilder.newBlockBuilderLike(null, selectedPositions.size())).collect(toImmutableList());
> ```



> newBlockBuilderLike

@mbasmanova Thanks for reviewing. Although there is only one caller now, I'm planning to add `uncheckedWriteXXX()` and `uncheckedAppendStructure()` etc. to the BlockBuilder interface. Note that the `reserve()` currently only reserves the arrays(offsets, nulls) for the current level of BlockBuilders while not the nested ones. E.g. calling it on ArrayBlockBuilder won't reserve the element BlockBuilder's internal arrays. This is because at the time of calling `reserve` on the ArrayBlockBuilder, we don't know how many elements will be inserted into the element BlockBuilder. Thus the element BlockBuilders would still suffer from calling `ensureCapacity()` for each inserted value. The planed `uncheckedAppendStructure()` for these types will need to call `reserve` on the element BlockBuilders when it's appending a new top level row, so that the element BlockBuilders won't need to grow memory on the fly.

I can add a method `newBlockBuilderLike(null, selectedPositions.size())` which calls `reserve()`. But the fact `reserve()` being there would make the mentioned future PRs a lot easier. What do you think?