@mbasmanova 2020-09-29
@yingsu00 newBlockBuilderLike internally computes the expectedEntries and creates a new builder that pre-allocates that many entries. Passing expected number of entries to newBlockBuilderLike makes more sense, then calling it with no hint, then immediately calling `reserve`. Hence, for the use case you have in this PR, `reserve` is not the right method to use.

If `reserve` or similar method becomes necessary in the future, it can be added on first use.