@yingsu00 2020-09-29
> @yingsu00 newBlockBuilderLike internally computes the expectedEntries and creates a new builder that pre-allocates that many entries. 

@mbasmanova  Actually, `newBlockBuilderLike` doesn't pre-allocates any memory but just set the `initialEntryCount`:

```
@Override
    public BlockBuilder newBlockBuilderLike(BlockBuilderStatus blockBuilderStatus)
    {
        return new Int128ArrayBlockBuilder(blockBuilderStatus, calculateBlockResetSize(positionCount));
    }

public Int128ArrayBlockBuilder(@Nullable BlockBuilderStatus blockBuilderStatus, int expectedEntries)
    {
        this.blockBuilderStatus = blockBuilderStatus;
        this.initialEntryCount = max(expectedEntries, 1);

        updateDataSize();
    }
```
It's up to the growCapacity() to allocate memory, but when it's uninitialized it just grow to `valueIsNull.length` which was just 1. So then whenever a new value got added, growCapacity() just grow array by 50% a time.

```
private void growCapacity()
    {
        int newSize;
        if (initialized) {
            newSize = BlockUtil.calculateNewArraySize(valueIsNull.length);
        }
        else {
            newSize = initialEntryCount;
            initialized = true;
        }

        valueIsNull = Arrays.copyOf(valueIsNull, newSize);
        values = Arrays.copyOf(values, newSize * 2);
        updateDataSize();
    }
```

If we want to pre-allocate memory, we need to do it explicitly. That's why we need `reserve()`. So we have three choices here:
1. Have reserve() in the BlockBuilders, and call it in the constructor:
```
public Int128ArrayBlockBuilder(@Nullable BlockBuilderStatus blockBuilderStatus, int expectedEntries)
    {
        this.blockBuilderStatus = blockBuilderStatus;
        this.initialEntryCount = max(expectedEntries, 1);
        
        reserve(expectedEntries);

        updateDataSize();
    }
```
Or
2. Change growCapacity() to
```
private void growCapacity()
    {
        int newSize = max( BlockUtil.calculateNewArraySize(valueIsNull.length), initialEntryCount);
        initialized = true;

        valueIsNull = Arrays.copyOf(valueIsNull, newSize);
        values = Arrays.copyOf(values, newSize * 2);
        updateDataSize();
    }
```
3. The third option is what this PR does. It can be changed that we add a new `public Int128ArrayBlockBuilder(@Nullable BlockBuilderStatus blockBuilderStatus, int expectedEntries, boolean reserveMemory)` and/or a new `newBlockBuilderLike(BlockBuilderStatus blockBuilderStatus, int expectedEntries)`. 
```
public Int128ArrayBlockBuilder(@Nullable BlockBuilderStatus blockBuilderStatus, int expectedEntries, boolean reserveMemory)
    {
        this.blockBuilderStatus = blockBuilderStatus;
        this.initialEntryCount = max(expectedEntries, 1);

        //------------------------
        if (reserveMemory) {
            reserve(expectedEntries);
        }
        //--------------------

        updateDataSize();
    }

public BlockBuilder newBlockBuilderLike(BlockBuilderStatus blockBuilderStatus, int expectedEntries)
    {
        BlockBuilder blockBuilder = new Int128ArrayBlockBuilder(blockBuilderStatus, expectedEntries, true);
    }
```
Then we can make the projection call this new constructor:
```
blockBuilders = blockBuilders.stream().map(blockBuilder -> blockBuilder.newBlockBuilderLike(null, selectedPositions.size())).collect(toImmutableList());
```

Yes true we can make these new constructors or `newBlockBuilderLike` just do what reserve() does but I think it's better to encapsulate the reserve logic in `reserve()` function for better readability and extensibility.

Both 1 and 2 would be affecting global usages. For many of the cases, we don't really know how big the BlockBuilders will be growing into. It's not very safe to do them that way. We could be seeing many BlockBuilders reserved more than the required. But for this case(projection) we do know how many elements are going to be added so it's safer to apply it just for projection for now. I tend to pick 3. What do you think?