@mbasmanova 2020-09-29
@yingsu00 I could be misreading the code, but here is what I understand. Int128ArrayBlockBuilder constructor doesn't allocate memory indeed, but it stores expectedEntries in initialEntryCount which is used to allocate memory in growCapacity whenever first value is added. In this case initialized is false, and growCapacity allocates enough memory to fit initialEntryCount.