@yingsu00 2020-09-30
@mbasmanova Thanks for reviewing the PR and bringing up the suggestion. I read the `growCapacity()` code wrong and you can disregard my last comment. I updated the PR by adding a new `newBlockBuilderLike(BlockBuilderStatus blockBuilderStatus, int expectedEntries)`. The performance benchmark shows similar improvement as before. Will you please take another look?