@yingsu00 2020-09-30
@mbasmanova Thank you so much for the rewrite! Yes it works. I just didn't use Type.createBlockBuilder much. I have updated the PR using the way you suggested.