@mbasmanova 2020-09-30
@yingsu00 Thank you. I have just one more question. My read of the existing com.facebook.presto.operator.project.GeneratedPageProjection#project implementation is that first page will be processed slowly as memory will be allocated incrementally. However, second and subsequent pages will start with block builders that have expectedEntries set to the number of rows processed in the previous page. If the number of rows being processed doesn't vary much from page to page, we shouldn't see any issues after the first page. Is it the case that we are seeing memory allocations because the number of rows in pages goes up and down a lot?